Fixed player and enemy being stuck in walls
Added a function CheckSidesWalk() to both the enemy and player controllers which checks if there is a wall within a certain distance of the unit before allowing them to move. If there is a wall within said distance in their movement direction, they will not be able to walk that way.